TwoStateAnts, Agent.act() takes clock time, fixes
Added a new application, TwoStateAnts.  TwoStateAnts have two modes:
in the first they avoid walls and other ants just like the tutorial, in the
second they approach ants to "visit" with them. After visiting or avoiding
for a set period of time, the ants switch modes. This is intended to be an
example of a multi-state behavior that a single state model can't learn.

In order to allow the agent to perform behaviors for a fixed period of "real"
time, independent of the rate of the simulation, they can't operate on "steps".
So the time value passed in to the act method got changed to a double, which
should be computed by the Body as the current wall clock time of the simulation
(simulation_steps*simulation_resolution).

Fixed BTFData which was failing to parse the final line of a btf file, leading
to off-by-one errors. Also modified AddSensorsBTF to reduce the chance of
catastrophic failure if the BTF files aren't perfectly aligned.