Fixed AKDE error. Speed improvements.
Typo in the code was causing the adaptive KDE function to return incorrect
estimates. Also memoized the k-nearest neighbor lookup so that the distance
calculation is only computed after the samples have been updated (repeated
calls to estimate do not cause a kd-tree look-up unless the tree changes).

Some quick testing shows that the non-adaptive KDE might be easier to tune
in situations where there are many samples.  This agrees with current research
in that adaptive KDE only improves estimates in parts of the space with
fewer samples.