Debugging
Split out the data parsing used in BIOHMM into a new class to make
it easier to apply BIOHMM learning to different things. BIOHMMInputParser
is the base class, SimpleInputParser is a derived class for testing learning
on output generated by TestGen. TestGen generates simple output that should be
easier to analyze.