Debugging
I've decided that optimizing the partition is the wrong way of
doing things.  Removing samples from one or another KDE can
decrease the likelihood of a sequence, even if we partition the
data according to the most likely sequence of states, or the
sequence of states that maximizes the overal likelihood of the
data. SO! Now, when we update the output function, we only ADD
samples according to our current estimate of the most likely
sequence. In order to keep from overwhelming the KDE, we now using
weighting, so that if we've already got a point recorded, we
increment the corresponding weight instead of adding a new point.
Since this changes how we should do our KNN (as a weighted KNN),
I've added the sample points and their weights to the parameter
file output.  Good news: currently I haven't encountered ANY
decrease in the loglike yet, so I think I've cracked the actual
algorithm.  Bad news: converged models still look like random
noise even for the toy test problem. Also, still have to think
about the actual probability distribution for the output.