Debugging - getting close now
Found another bug, this one in how xi was getting calculated. After
some post-bug-fix testing, it looks like it's working in the
following sense: 1) when loglike decreases, it's by small amounts
which is consistant with the other numerical instability that we
see within an iteration, and 2) the transition matrix is being
moved closer to the true matrix iteration to iteration even when
LL goes down. Tested first by fixing the initial parameters to
the correct values (transitionfunction, prior, initial kde).
Then let the transition function start at uniform, then removed
the pre-seeded kde for original kde seeding.  Am pushing this
update to mellifera to test for a longer period of time to see
where it comes out.