Debugging
It appears that weighting the sample added to the output function
for each state by the probability the sequence was in that state
at the time it observed the output does not converge as quickly as
simply improving the weight of each sample for the state in the
viterbi most likely sequence.  Test gen updated to have non
singular transition matrix, SimpleInputParser updated to work with
current BIOHMM class.