New TwoStateAnt
I've changed the twostateants package because of the issues with
learning the time-only avoid/approach ants. TwoStateAnt now runs
a simple single-source foraging simulation where the ants
alternate between visiting the "nest" poi and a "food" poi. This
update includes changes to the btf loggers, the BIOHMM input
parser, three new sensor interfaces, and the body implementations
for those sensors.