Linear regression fish, and improved graphics
Added new biosim app for using linear regression on the fish
dataset to learn a single state behavior. Also added new
FishPortrayal which uses a png icon. Useful as an example
of using an image file as a body portrayal. Currently very slow,
might be possible to speed up by caching part of the
AffineTransform.