Environmental properties as a config file
No more recompiling to change params! See SmallDomWorld.java for
an example.  Made small changes to Environment to allow setting
the seed when using newSimulation(...). Before, it was assumed
that when using the GUI you didn't care about the seed.

Also modified DomWorldStateMachine.java so that monkeys don't
react to monkeys in their personal space if those monkeys are
chasing or being chased. This prevents some "chain reactions".