Config, heirarchy stability, encounter hysteresis
Relocated some of the static config to the DomWorldStateMachine
class.

Added a parameter for heirarchy stability so that
close encounters do not always lead to fights.

Added hysteresis to the "tooClose" method, so that if we decide
not to fight because the heirarchy is stable, we don't repeatedly
check if we should fight for as long as the close monkey is
within the fight radius.

Put another way, we only check if we want to fight the first time
a monkey enters the PERSONAL_DIST radius.

If another monkey enters PERSONAL_DIST, but is not the closest,
we still do not initiate a new check (potential error).