Bugfix: getSelfVelXYT()
When the API change happened, the old data members that were being used
to hold on to the body velocity from previous steps were being shadowed.
This change should fix that. This bug didn't show up in any of the included
applications since none of them were using this velocity information.