Added arbitrary IDs and ID parsing for logging
Simulation now keeps track of a logical ID for each body in the
sim. This map isn't garunteed to exist for everything, so if it
fails, it uses the agent's index in the list as the ID (old
behavior). No garuntees that the IDs are unique, so be careful
you don't give two bodies the same ID.