Changed labeling
So... keeping a label independent of the Body object is dumb.
For one, we already have tools from MASON to inspect and track it.
So now all Body objects have a String "label" data member. In the
Environment class, this is set as the index of when the body was
added. In the InitiallyPlacedEnvironment class, the addBody(...)
method is  overloaded, and the set of parsed IDs is used. Parsed
IDs are set to incoming Body objects in the order that the Body
objects are added, until no more parsed IDs are left. Then, we
start assigning from the positive integers, starting with one
plus the last parsed ID, making sure to not re-use any previously
assigned IDs by checking the usedIDs HashSet. NOTE: This doesn't
garuntee that the parsed IDs are always less than the generated
IDs, and it doesn't garuntee that the generated IDs will not
match any IDs that show up after the first time step in the
parsed BTF. Use with caution, YMMV, you've been warned!