FishReynolds combines FishLR and FishKNN
In an effort to clean up the two projects, preserve working demos,
and make the interface with data-as-demonstrator code cleaner,
FishReynolds combines FishLR and FishKNN into one class. This is
the first commit and may contain subtle bugs, even though this
was tested briefly. Changes to C code wrapper make files are
platform specific and may have to be reverted for your platform.