Speed ups and clean ups
Instead of rebuilding every query, allow flann to decide when to
rebuild by using addPoint(...). This requires keeping track of
points as they are added.

NOTE: It is unclear if flann::Index copy constructor creates
memory leaks (it doesn't appear so at the moment), or is thread
safe (it appears safe).