The great java properties migration part 1
In order to standardize the arguments on the command line between
simulation for visualization purposes and learning, and to keep a
record of what the actual settings were for everything, I've
started swapping everything over to java properties.

IMPORTANT NOTE!
The philosophy for java properties objects with settings is the
following: Properties objects are used to configure objects, or
to report configured settings, but they are transient.
Specifically they should NOT be kept around by the objects that
use them if possible. Data members or getter-setter methods
specific to each class should hold data, not generic Properties
objects.

- Learners now have configure(Properties p), and getSettings()
  methods.
- DataAsDemonstrator.java now takes command line arguments in java
  properties format (space separated key,value pairs) using the
  new ArgsToProps.java class. The command line properties are
  passed to ReynoldsFeatures, and used as defaults for the
  settings passed to the LearnerAgent. One of the command line
  properties ("--learnerSettings") can also specify a seperate
  property file for the learner (to avoid cluttering the command
  line).
- ReynoldsFeatures.java now also grabs the settings from a
  configured learner and includes this in the Properties object
  returned by getSettings(). This means the "settings.prop" file
  generated by DataAsDemonstrator.java also includes all of the
  learner settings as well.
- Subsequent commits will transition FishReynolds.java to use the
  same format for command line arguments