Forgot to scale max theta velocity
Theta velocity is bounded due to pre-processing: On consecutive
frames, if the theta difference is greater than Pi, we assume the
rotation took the smaller arc. This means the greatest observable
theta velocity is Pi radians per frame, which is 30*Pi radians per
second. The max allowed theta velocity in the fish body was set
to 2*Math.PI radians per seconds, much smaller. This not only
effects Data-as-Demonstrator (as "corrected" samples would likely
be outside this range), but also the performance of normal models
in terms of x-vel histograms