FastSlowFish.java added
FastSlowFish implements two mode schooling behavior. The basic
behavior is RBF-Reynold's with fixed gains. Every time step each
fish has a small probability of entering a "go fast" state, where
it's desired forward velocity will be scaled up by some constant
factor. The fish remains in this state for a short fixed period.
If the fish is not currently in the fast state, it's forward
velocity is scaled according to the following formula:

 scale = baseScale + maxScaleMultiplier*max(0, tanh(gain*(xvmean-baseSpeed)))

This will increase the fish's base speed (up to a saturation
threshold) if it is slower than the average velocity of other fish
in the school.

The reasoning behind this was to try and create a simulation that
matched the observed asymmetric exponential tail of the
distribution of real fish x-velocity. Unfortunately, the
distribution of the x-velocity of this simulation is actually
bi-modal, and is a poor fit for the observed distribution.